Function jemallocator::ffi::mallctlnametomib

pub unsafe extern "C" fn mallctlnametomib(
    name: *const i8, 
    mibp: *mut usize, 
    miblenp: *mut usize
) -> i32

Translates a name to a “Management Information Base” (MIB) that can be passed repeatedly to mallctlbymib.

This avoids repeated name lookups for applications that repeatedly query the same portion of the namespace.

On success, mibp contains an array of *miblenp integers, where *miblenp is the lesser of the number of components in name and the input value of *miblenp. Thus it is possible to pass a *miblenp that is smaller than the number of period-separated name components, which results in a partial MIB that can be used as the basis for constructing a complete MIB. For name components that are integers (e.g. the 2 in arenas.bin.2.size), the corresponding MIB component will always be that integer.