Type Alias jemallocator::ffi::extent_decommit_t

pub type extent_decommit_t = unsafe extern "C" fn(_: *mut extent_hooks_s, _: *mut c_void, _: usize, _: usize, _: usize, _: u32) -> bool;

Extent decommit function.

Decommits any physical memory that is backing pages within an extent at given addr and size at offset bytes, extending for length on behalf of arena arena_ind, returning false upon success, in which case the pages will be committed via the extent commit function before being reused.

If the function returns true, this indicates opt-out from decommit; the memory remains committed and available for future use, in which case it will be automatically retained for later reuse.